# target/classes/api.yml
openapi: 3.0.3
info:
  title: Credit Service API
  version: 4.0.0
  description: >
    Credit products API.
    - Loans (PERSONAL/BUSINESS) live under /credits
    - Credit card lines live under /credit-lines
servers:
  - url: http://localhost:8585
...
paths:
...
  /credits:
    post:
      tags: [Loans]
      summary: Create loan
      description: >
        Creates a PERSONAL or BUSINESS loan.
         Rules:
         - PERSONAL: only one active loan per customer.
         - BUSINESS: multiple loans allowed.
        - Admission: reject if customer has overdue debts.
      operationId: registerLoan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoanRequest' }
      responses:
        '201':
          description:  Loan created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

    get:
      tags: [ Loans ]
      summary: List loans
      description: >
        Returns loans. Use filters to scope the result. # Listado global con filtros
      operationId: listLoans
      parameters:
        - in: query
          name: customerId
          schema: { type: string }
          description: Filter by customer id.
        - in: query
          name: type
          schema: { $ref: '#/components/schemas/LoanType' }
          description: Filter by loan type.
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/CreditStatus' }
          description: Filter by status.
        - in: query
          name: includeClosed
          schema: { type: boolean, default: false }
          description:  Include CLOSED loans.
      responses:
        '200':
          description: Loan list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LoanResponse' }

#  /customers/{customerId}/credits:
#    get:
#      tags: [Loans]
#      summary: List credits by customer
#      description: Return loans of the client specific.
#      operationId: listLoansByCustomer
#      parameters:
#        - in: path
#          name: customerId
#          required: true
#          schema: { type: string }
#          description: Customer identifier.
#              - in: query
#          name: type
#          schema: { $ref: '#/components/schemas/CreditType' }
#          description: Filtra por tipo de crédito.
#        - in: query
#          name: status
#          schema: { $ref: '#/components/schemas/CreditStatus' }
#          description: Filtra por estado del crédito.
#      responses:
#        '200':
#          description: List of client credits
#          content:
#            application/json:
#              schema:
#                type: array
#                items: { $ref: '#/components/schemas/LoanResponse' }
#        '404': { $ref: '#/components/responses/NotFound' }
...
  /credits/overdue/{customerId}:
    get:
      tags: [ Common ]
      summary: Customer overdue status
      description: Returns consolidated overdue info across all credit products. # Consolida cartera
      operationId: getDebtStatusByCustomer
      parameters:
        - in: path
          name: customerId
          required: true
          schema: { type: string }
          description: Customer identifier.
      responses:
        '200':
          description: Customer Debt Summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OverdueStatus' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  responses:
    BadRequest:
      description: Validation or bad request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict with business rules or resource state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnprocessableEntity:
      description: Semantically valid but violates domain constraints
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    LoanType:
      type: string
      description: Loan product type.
      enum: [ PERSONAL, BUSINESS ]  # Tipos de PRÉSTAMO
    LoanResponse:
      type: object
      description: Loan stored in the system.
      required: [ id, customerId, type, limit, balance, interestAnnual, status, createdAt ]
      properties:
        id: {
          type: string,
          description: Loan id,
          readOnly: true
        }
        customerId: {
          type: string,
          readOnly: true
        }
        type: {
          $ref: '#/components/schemas/LoanType',
          readOnly: true
        }
        limit: {
          type: number,
          minimum: 0,
          description: Approved amount (max).
        }
        balance: {
          type: number, minimum: 0, readOnly: true
        }
        interestAnnual: {
          type: number,
          minimum: 0
        }
        dueDate:
          type: string
          format: date
          description: Next payment due date. # Fecha de pago próxima
        status: {
          $ref: '#/components/schemas/CreditStatus',
          readOnly: true
        }
        closedAt: {
          type: string,
          format: date-time,
          readOnly: true
        }
        closeReason: {
          type: string,
          readOnly: true
        }
        createdAt: {
          type: string,
          format: date-time,
          readOnly: true
        }
        updatedAt: {
          type: string,
          format: date-time,
          readOnly: true
        }
        currency: {
          type: string,
          example: PEN
        }
    LoanRequest:
      type: object
      description: Loan creation payload. # Payload de creación de préstamo
      additionalProperties: false
      properties:
        customerId: { type: string, description: Internal customer id. }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        documentNumber: { type: string }
        type: { $ref: '#/components/schemas/LoanType' }
        limit: { type: number, minimum: 0 }
        interestAnnual: { type: number, minimum: 0 }
        dueDate:
          type: string
          format: date
          description: Next payment due date.
        currency:
          type: string
          example: PEN
          default: PEN
      required: [ type, limit, interestAnnual ]
      oneOf:
        - required: [ customerId ]
        - required: [ documentType, documentNumber ]
...
    CreditStatus:
      type: string
      description: Estado del crédito.
      enum: [ ACTIVE, CLOSED, DELINQUENT ]
    DocumentType:
      type: string
      enum: [ DNI, RUC, CE ]
...
    OverdueStatus:
      type: object
      required: [ hasOverdue, totalOverdueAmount ]
      properties:
        hasOverdue:
          type: boolean
          description: True if at least one product is DELINQUENT. # Tiene mora
        totalOverdueAmount:
          type: number
          description: Sum of balances for all DELINQUENT products.
...
    Error:
      type: object
      description: RFC 7807 Problem Details. # Estructura de error
      additionalProperties: false
      properties:
        type: { type: string, example: 'https://errors.bank.com/business-rule' }
        title: { type: string, example: 'Business rule violation' }
        status: { type: integer, example: 409 }
        detail: { type: string, example: 'Customer already has an active PERSONAL loan.' }
        instance: { type: string, example: '/credits/12345' }
        code: { $ref: '#/components/schemas/ErrorCode' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      description: Field validation error. # Error por campo
      additionalProperties: false
      properties:
        field: { type: string, example: 'interestAnnual' }
        message: { type: string, example: 'Must be greater or equal than 0' }
        code: { type: string, example: 'Min' }
    ErrorCode:
      type: string
      description: Business code. # Código interno
      enum:
        - CREDIT_PERSONAL_LIMIT
        - CREDIT_DELINQUENT_LOCK
        - CREDIT_CARD_LIMIT_EXCEEDED
        - CREDIT_INVALID_STATE
        - CREDIT_NOT_FOUND
        - CUSTOMER_NOT_FOUND
        - VALIDATION_ERROR
        - SERVICE_UNAVAILABLE

  # Credit Lines
...