# target/generated-sources/openapi/src/main/java/com/nttdata/credit/model/LoanType.java
package com.nttdata.credit.model;

import java.util.Objects;
import io.swagger.annotations.ApiModel;
import com.fasterxml.jackson.annotation.JsonValue;
import org.openapitools.jackson.nullable.JsonNullable;
import javax.validation.Valid;
import javax.validation.constraints.*;

import com.fasterxml.jackson.annotation.JsonCreator;

/**
 * Loan product type.
 */
public enum LoanType {
  
  PERSONAL("PERSONAL"),
  
  BUSINESS("BUSINESS");

  private String value;

  LoanType(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static LoanType fromValue(String value) {
    for (LoanType b : LoanType.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}